refactor(mcp): move in-process FastMCP adapter out of Academy
